--- README.md
@@ -1,2 +1,0 @@
-# MassBalanceMachine
-Global machine learning glacier mass balance modelling, capable of assimilating all sources of glaciological and remote sensing data
