--- data_processing/data/files/Iceland_Stake_Data_Nan_Glaciers.csv
@@ -1,6 +1,0 @@
-stake,yr,d1,d2,d3,lat,lon,elevation,rhow,rhos,bw_stratigraphic,bs_stratigraphic,ba_stratigraphic,bw_floating_date,bs_floating_date,ba_floating_date,RGIId,GLIMSId,Name
-L01,1998,12/10/1997,17/04/1998,16/09/1998,64.49278,-20.43633,490.0,490.0,,0.29,-7.32,-7.03,0.29,-7.32,-7.03,,,
-L01,1999,16/09/1998,22/04/1999,28/09/1999,64.49382,-20.43648,505.0,547.0,,1.669,-6.619,-4.95,1.669,-6.619,-4.95,,,
-L01,2000,28/09/1999,27/04/2000,04/10/2000,64.49343,-20.43622,490.0,508.0,,1.32,-6.855,-5.535,1.32,-6.8025,-5.4825,,,
-L01,2001,04/10/2000,19/04/2001,05/10/2001,64.49382,-20.43647,485.0,591.194968553459,,0.937,-7.246,-6.309,0.8845,-7.246,-6.3615,,,
-B08,1995,15/09/1994,03/05/1995,,64.7766607457982,-16.0565003060087,715.0,613.636363636364,,0.54,,,0.54,,,,,
