--- data_processing/data/stakes/HOFS.txt
@@ -1,1 +1,0 @@
-hn9v, hn10v, hn12v, hn9, hn10, hn12, hn13, hn14, hn15, hn16, hn17, hn12a, hn13a, hn14a, hn15a, hn16a, hn14aa, hn15aa, hn16aa, h18, hsa7, hsa7v, hsa8, hsa9, hsa10, hsa11, hsa13, hsa15, hsa16, hsa17, hsv8, hsv9, hsv11, hsv13, hsv15, hsv17, blt8, blt11
