--- data_processing/data/stakes/LANG.txt
@@ -1,1 +1,0 @@
-L01, L02, L03, L04, L05, L06, L07, L08, L09, L10, L11, L12, L13, L14, L15, L16, L17, L18, L19, L20, L21, L22, L23, L24, L25
